refactor(convex): one corpus loader per adapter, not one per suite
Follow-up to #380. `src/corpus.ts` arrived with the translator unit test and
left `src/adversarial.test.ts` holding a second, independent reader of the same
corpus files: its own `ActionsFile` interfaces, its own `requireMessage`, its
own ORACLE_ACTIONS / THROWING_ACTIONS derivation, and its own validators. Two
answers to "which shapes must this adapter refuse" is one too many. Mongoose is
the precedent — `mongoose/src/adversarial.test.ts` reads its classification and
mapper from `mongoose/src/corpus.ts`.

Cross-adapter duplication is untouched and stays deliberate (ADR 0007); this is
only about the two copies inside convex.

Moved into `src/corpus.ts`:

- the JSON validators and `assertKeys`, and `parseActionsFile`, which replaces
  an `as ActionsFile` cast in the translator test with a real parse. The
  `adapterUnsupported` / `adapterSupportedExpected` / `knownDivergences` fields
  become required, since the corpus always carries them and validating them is
  what makes that true rather than assumed.
- the seeds and derived-fields readers, with the declared-key guards inside
  them: the seed and tag key sets, the principal and its attribute shapes, and
  the derived-fields roster. They are exposed as `parseSeedsFile` /
  `parseDerivedFile` rather than run on import, so the offline suite neither
  pays for them nor trips over a guard about data it never touches.
- `nullRepresentationOmittedFor`, which both suites needed and both derived.

`ThrowingAction` becomes an object rather than the tuple #380 copied from
drizzle: the harness already destructures `{action, message}` and titles its
cases with `$action`, and carrying `reason` gives the translator test a better
title than a positional `%s`.

This is close to line-neutral, and that is honest: the genuine duplicate was
~150 lines (the interfaces, `requireMessage`, the classification). The rest —
the runtime validators and the key guards — moved rather than collapsed,
because `corpus.ts` never had them.

Verified: `npm run typecheck`, `npm test` (457 offline), `npm run
test:integration` (25), and the adversarial suite against the pinned PDP
(218/218, self-hosted Convex backend). Each of the five declared-key guards was
re-proved from its new home by perturbing the corpus and watching the suite
fail at load: an extra seed key, an extra principal attribute, a reshaped
principal attribute, an extra derived field, and a missing derived entry. The
corpus was restored after each.
