+++ convex/src/corpus.ts
@@ -9,6 +9,7 @@ import {
 import type {
   PlanExpressionOperand,
   PlanResourcesResponse,
+  Principal,
   Value,
 } from "@cerbos/core";
 
@@ -29,6 +30,11 @@ import { PlanKind } from ".";
  *   definition, three readers, no copy to drift.
  * - **the classification.** Which actions this adapter must refuse, and with which message, is a
  *   corpus decision (`actions.json`), not a per-suite one.
+ *
+ * The declared-key guards live here for the same reason, even though only the harness consumes
+ * seeds: they are one adapter's statement of what it reads out of the corpus, and a second copy
+ * would be a second answer. They are exposed as `parse*` functions rather than run on import, so
+ * the offline suite neither pays for them nor trips over a guard about data it never touches.
  *
  * The code in this file is duplicated across adapters **on purpose** — adapters share data, not
  * code, so that every adapter stays standalone. Do not extract it into `conformance/`, do not
@@ -56,6 +62,55 @@ export function readCorpusJson(file: string): unknown {
   return JSON.parse(fs.readFileSync(path.join(CONFORMANCE_DIR, file), "utf8"));
 }
 
+// -- corpus JSON validators ------------------------------------------------------------------------
+//
+// A corpus file is machine-generated and diffed by validate-corpus.sh, so these are not here to
+// catch a malformed file. They are what stops a field this repository RENAMED from arriving as
+// `undefined` and being read as "absent" by a suite that would then assert nothing — the projection
+// trap `conformance/README.md` describes, applied to the parse rather than to a hand-written
+// projection.
+
+export const isRecord = (value: unknown): value is Record<string, unknown> =>
+  typeof value === "object" && value !== null && !Array.isArray(value);
+
+export const isStringArray = (value: unknown): value is string[] =>
+  Array.isArray(value) && value.every((item) => typeof item === "string");
+
+const isMessageMap = (value: unknown): value is Record<string, string> =>
+  isRecord(value) &&
+  Object.values(value).every((message) => typeof message === "string");
+
+/**
+ * Set equality between the keys a corpus record carries and the keys a suite declares it consumes.
+ *
+ * Both directions matter. A corpus key nothing reads is dropped from the stored document and the
+ * `check()` oracle at once, so the differential agrees for the wrong reason; a declared key the
+ * corpus no longer carries is a suite reading `undefined` and calling it data.
+ */
+export function assertKeys(
+  label: string,
+  got: string[],
+  want: readonly string[],
+  optional: readonly string[] = [],
+): void {
+  const allowed = new Set<string>([...want, ...optional]);
+  for (const key of got) {
+    if (!allowed.has(key)) {
+      throw new Error(
+        `${label} carries "${key}", which this harness does not consume: an unconsumed corpus field is dropped from the stored document and the check() oracle at once`,
+      );
+    }
+  }
+  const present = new Set(got);
+  for (const key of want) {
+    if (!present.has(key)) {
+      throw new Error(
+        `${label} is missing "${key}", which this harness consumes`,
+      );
+    }
+  }
+}
+
 // -- actions.json --------------------------------------------------------------------------------
 
 export interface UnsupportedShape {
@@ -65,7 +120,7 @@ export interface UnsupportedShape {
   messages: Record<string, string>;
 }
 
-export interface AdapterUnsupportedEntry {
+export interface AdapterOutcome {
   action: string;
   reason: string;
   /** Absent on `adapterSupportedExpected` / `nullRepresentationOmitted`, required on a throw. */
@@ -90,11 +145,71 @@ export interface KnownDivergence {
 
 export interface ActionsFile {
   conformance: string[];
-  adapterUnsupported?: Record<string, AdapterUnsupportedEntry[]>;
-  adapterSupportedExpected?: Record<string, AdapterUnsupportedEntry[]>;
+  adapterUnsupported: Record<string, AdapterOutcome[]>;
+  adapterSupportedExpected: Record<string, AdapterOutcome[]>;
   expectedUnsupported: UnsupportedShape[];
   nullRepresentationOmitted: NullRepresentationOmittedEntry[];
-  knownDivergences?: KnownDivergence[];
+  knownDivergences: KnownDivergence[];
+}
+
+const isUnsupportedShape = (value: unknown): value is UnsupportedShape =>
+  isRecord(value) &&
+  typeof value["action"] === "string" &&
+  typeof value["shape"] === "string" &&
+  isMessageMap(value["messages"]);
+
+const isAdapterOutcome = (value: unknown): value is AdapterOutcome =>
+  isRecord(value) &&
+  typeof value["action"] === "string" &&
+  typeof value["reason"] === "string" &&
+  // `adapterUnsupported` carries this and the classification below requires it;
+  // `adapterSupportedExpected` and `nullRepresentationOmitted` do not throw, so they do not.
+  (value["message"] === undefined || typeof value["message"] === "string");
+
+const isAdapterMap = (
+  value: unknown,
+): value is Record<string, AdapterOutcome[]> =>
+  isRecord(value) &&
+  Object.values(value).every(
+    (entries) => Array.isArray(entries) && entries.every(isAdapterOutcome),
+  );
+
+const isNullRepresentationOmitted = (
+  value: unknown,
+): value is NullRepresentationOmittedEntry =>
+  isRecord(value) &&
+  typeof value["action"] === "string" &&
+  typeof value["reason"] === "string" &&
+  isMessageMap(value["messages"]);
+
+const isKnownDivergence = (value: unknown): value is KnownDivergence =>
+  isRecord(value) &&
+  typeof value["action"] === "string" &&
+  isStringArray(value["adapters"]);
+
+/**
+ * `actions.json`, validated rather than cast.
+ *
+ * Every group the interface declares is checked: a group this predicate does not name would be
+ * read as `undefined` by whichever suite consumes it, and a suite that silently consumes nothing
+ * is the failure the corpus README warns about.
+ */
+export function parseActionsFile(value: unknown): ActionsFile {
+  if (
+    !isRecord(value) ||
+    !isStringArray(value["conformance"]) ||
+    !isAdapterMap(value["adapterUnsupported"]) ||
+    !isAdapterMap(value["adapterSupportedExpected"]) ||
+    !Array.isArray(value["expectedUnsupported"]) ||
+    !value["expectedUnsupported"].every(isUnsupportedShape) ||
+    !Array.isArray(value["nullRepresentationOmitted"]) ||
+    !value["nullRepresentationOmitted"].every(isNullRepresentationOmitted) ||
+    !Array.isArray(value["knownDivergences"]) ||
+    !value["knownDivergences"].every(isKnownDivergence)
+  ) {
+    throw new Error("Invalid conformance actions");
+  }
+  return value as unknown as ActionsFile;
 }
 
 /**
@@ -104,11 +219,12 @@ export interface ActionsFile {
  * mapper typo or an unrelated validation satisfies the assertion just as well as the limitation
  * the corpus documents (cerbos/query-plan-adapters#326).
  */
-export type ThrowingAction = readonly [
-  action: string,
-  reason: string,
-  message: string,
-];
+export interface ThrowingAction {
+  action: string;
+  /** Why the corpus says this adapter cannot express the shape. Titles read better with it. */
+  reason: string;
+  message: string;
+}
 
 export interface ActionClassification {
   oracleActions: string[];
@@ -133,10 +249,10 @@ export function classifyActionsForAdapter(
   manifest: ActionsFile,
   adapter: string,
 ): ActionClassification {
-  const unsupported = manifest.adapterUnsupported?.[adapter] ?? [];
+  const unsupported = manifest.adapterUnsupported[adapter] ?? [];
   const unsupportedActions = new Set(unsupported.map((entry) => entry.action));
   const supportedExpected = new Set(
-    (manifest.adapterSupportedExpected?.[adapter] ?? []).map(
+    (manifest.adapterSupportedExpected[adapter] ?? []).map(
       (entry) => entry.action,
     ),
   );
@@ -145,33 +261,278 @@ export function classifyActionsForAdapter(
     ...supportedExpected,
   ];
   const throwingActions: ThrowingAction[] = [
-    ...unsupported.map((entry): ThrowingAction => [
-      entry.action,
-      entry.reason,
-      requireMessage(
-        `adapterUnsupported.${adapter}.${entry.action}`,
-        entry.message,
-      ),
-    ]),
+    ...unsupported.map(
+      (entry): ThrowingAction => ({
+        action: entry.action,
+        reason: entry.reason,
+        message: requireMessage(
+          `adapterUnsupported.${adapter}.${entry.action}`,
+          entry.message,
+        ),
+      }),
+    ),
     ...manifest.expectedUnsupported
       .filter((entry) => !supportedExpected.has(entry.action))
-      .map((entry): ThrowingAction => [
-        entry.action,
-        entry.shape,
-        requireMessage(
-          `expectedUnsupported.${entry.action}.messages.${adapter}`,
-          entry.messages?.[adapter],
-        ),
-      ]),
+      .map(
+        (entry): ThrowingAction => ({
+          action: entry.action,
+          reason: entry.shape,
+          message: requireMessage(
+            `expectedUnsupported.${entry.action}.messages.${adapter}`,
+            entry.messages[adapter],
+          ),
+        }),
+      ),
   ];
 
   return {
     oracleActions: [...new Set(oracleActions)].sort(),
-    throwingActions: throwingActions.sort(([left], [right]) =>
-      left.localeCompare(right),
+    throwingActions: throwingActions.sort((left, right) =>
+      left.action.localeCompare(right.action),
     ),
     supportedExpected,
   };
+}
+
+/**
+ * The `nullRepresentationOmitted` entries, each carrying the message THIS adapter must reject with.
+ *
+ * Separate from the classification above because these are not a throw the adapter chose: every
+ * adapter must reject them, the two NULL conventions being indistinguishable on the wire (#302).
+ */
+export function nullRepresentationOmittedFor(
+  manifest: ActionsFile,
+  adapter: string,
+): (NullRepresentationOmittedEntry & { message: string })[] {
+  return manifest.nullRepresentationOmitted.map((entry) => ({
+    ...entry,
+    message: requireMessage(
+      `nullRepresentationOmitted.${entry.action}.messages.${adapter}`,
+      entry.messages[adapter],
+    ),
+  }));
+}
+
+// -- seeds.json and derived-fields.json -----------------------------------------------------------
+//
+// Only the adversarial harness consumes these — the translator unit test asserts filters, not rows
+// — but they live here with the rest of the corpus reader, and they are FUNCTIONS rather than
+// module-level parses so that importing this file for the offline suite neither reads them nor
+// fails on a guard about data it does not touch.
+//
+// The same parsed seed feeds the stored document AND the check() oracle, so a corpus field the
+// harness does not consume is dropped from both sides at once and the differential agrees for the
+// wrong reason — the projection trap conformance/README.md describes for actions.json, applied to
+// the seeds.
+
+export interface Tag {
+  id: string;
+  name: string | null;
+}
+
+export interface Seed {
+  id: string;
+  aBool: boolean;
+  aString: string;
+  aNumber: number;
+  aOptionalString: string | null;
+  tags: Tag[];
+  subCategoryNames: string[];
+  /** The seed whose scalars this row's to-one `parent` carries; null for no parent. */
+  parentSeedId: string | null;
+}
+
+export interface SeedsFile {
+  principal: Principal;
+  resourceKind: string;
+  seeds: Seed[];
+}
+
+/** One seed's derived fields, exactly as conformance/derived-fields.json carries them. */
+export interface DerivedEntry {
+  createdBy: string;
+  aDouble: number | null;
+  createdAt: string | null;
+  scope: string | null;
+  labels: (string | null)[];
+}
+
+export interface DerivedFile {
+  fields: string[];
+  derived: Record<string, DerivedEntry>;
+}
+
+const SEED_KEYS = [
+  "id",
+  "aBool",
+  "aString",
+  "aNumber",
+  "aOptionalString",
+  "tags",
+  "subCategoryNames",
+  "parentSeedId",
+] as const;
+
+/** Corpus prose, never read by a harness: the one documented exclusion from SEED_KEYS. */
+const SEED_NOTE_KEY = "note";
+
+/** The one nested object array a seed carries. A key added inside an element is dropped from both
+ * sides of the differential just as silently as a top-level one, so it is guarded the same way. */
+const TAG_KEYS = ["id", "name"] as const;
+
+const DERIVED_KEYS = [
+  "createdBy",
+  "aDouble",
+  "createdAt",
+  "scope",
+  "labels",
+] as const;
+
+// The corpus principal is guarded the same way and for the same reason. It feeds the PLAN under
+// test AND the check() oracle, so an attribute dropped on the way in vanishes from both sides at
+// once: the plan folds to ALWAYS_DENIED and the oracle, built from the same principal, agrees. That
+// is how langchain-chromadb's hardcoded attribute allowlist let `pv-exists` pass while testing
+// nothing (conformance/README.md, "Adding a new hostile shape", step 7). The harness passes the
+// principal through verbatim, which is correct; the guard is what proves it still does.
+//
+// `id` and `roles` are deliberately IN scope, guarded by PRINCIPAL_KEYS one level above the
+// attributes — the same two-level shape SEED_KEYS and TAG_KEYS use for a row and its `tags[]`
+// elements. A role dropped on the way in changes every policy decision at once; that it is less
+// likely to be projected away than an attribute is a reason to expect the assertion to stay quiet,
+// not a reason to omit it.
+const PRINCIPAL_KEYS = ["id", "roles", "attr"] as const;
+
+const PRINCIPAL_ATTR_KEYS = [
+  "allowedTags",
+  "context",
+  "fewTeams",
+  "manyTeams",
+] as const;
+
+/**
+ * One principal attribute, checked against the two JSON shapes the corpus carries. A key-set guard
+ * says nothing about a change inside a value and three of the four attributes are lists, so the
+ * element type is asserted for the same reason the seed guard descends into `tags[]`.
+ */
+function assertPrincipalAttrShape(label: string, value: unknown): void {
+  if (typeof value === "string") return;
+  if (isStringArray(value)) return;
+  throw new Error(
+    `${label} is neither a string nor an array of strings, the only two shapes this harness consumes: a reshaped principal attribute feeds the plan and the check() oracle at once`,
+  );
+}
+
+const isTag = (value: unknown): value is Tag =>
+  isRecord(value) &&
+  typeof value["id"] === "string" &&
+  (typeof value["name"] === "string" || value["name"] === null);
+
+const isSeed = (value: unknown): value is Seed =>
+  isRecord(value) &&
+  typeof value["id"] === "string" &&
+  typeof value["aBool"] === "boolean" &&
+  typeof value["aString"] === "string" &&
+  typeof value["aNumber"] === "number" &&
+  (typeof value["aOptionalString"] === "string" ||
+    value["aOptionalString"] === null) &&
+  Array.isArray(value["tags"]) &&
+  value["tags"].every(isTag) &&
+  isStringArray(value["subCategoryNames"]) &&
+  (typeof value["parentSeedId"] === "string" || value["parentSeedId"] === null);
+
+const isPrincipal = (value: unknown): value is Principal =>
+  isRecord(value) &&
+  typeof value["id"] === "string" &&
+  isStringArray(value["roles"]) &&
+  isRecord(value["attr"]);
+
+const isDerivedEntry = (value: unknown): value is DerivedEntry =>
+  isRecord(value) &&
+  typeof value["createdBy"] === "string" &&
+  (typeof value["aDouble"] === "number" || value["aDouble"] === null) &&
+  (typeof value["createdAt"] === "string" || value["createdAt"] === null) &&
+  (typeof value["scope"] === "string" || value["scope"] === null) &&
+  Array.isArray(value["labels"]) &&
+  value["labels"].every((label) => label === null || typeof label === "string");
+
+/**
+ * `seeds.json`, validated and then held to the key sets above.
+ *
+ * The returned rows are the parsed JSON verbatim, never rebuilt field by field: `Object.keys` has
+ * to report the CORPUS key set for the guard to say anything, and a parser that copied each field
+ * it knew about could only ever report the keys already named here.
+ */
+export function parseSeedsFile(value: unknown): SeedsFile {
+  if (
+    !isRecord(value) ||
+    !isPrincipal(value["principal"]) ||
+    typeof value["resourceKind"] !== "string" ||
+    !Array.isArray(value["seeds"]) ||
+    !value["seeds"].every(isSeed)
+  ) {
+    throw new Error("Invalid conformance seeds");
+  }
+  const seedsFile = value as unknown as SeedsFile;
+
+  seedsFile.seeds.forEach((seed, index) => {
+    const label = `seeds.json seeds[${index}]`;
+    assertKeys(label, Object.keys(seed), SEED_KEYS, [SEED_NOTE_KEY]);
+    seed.tags.forEach((tag, tagIndex) => {
+      assertKeys(`${label}.tags[${tagIndex}]`, Object.keys(tag), TAG_KEYS);
+    });
+  });
+
+  assertKeys(
+    "seeds.json principal",
+    Object.keys(seedsFile.principal),
+    PRINCIPAL_KEYS,
+  );
+  // `attr` is optional on the SDK's Principal type; the corpus always carries it, and the
+  // assertion above is what proves it rather than this fallback.
+  const attr = seedsFile.principal.attr ?? {};
+  assertKeys("seeds.json principal.attr", Object.keys(attr), PRINCIPAL_ATTR_KEYS);
+  for (const [key, attrValue] of Object.entries(attr)) {
+    assertPrincipalAttrShape(`seeds.json principal.attr.${key}`, attrValue);
+  }
+
+  return seedsFile;
+}
+
+/**
+ * `derived-fields.json`, validated and held to `DERIVED_KEYS` — and to the seed roster, so a seed
+ * added without its derived entry fails here rather than at whichever action first reads it.
+ */
+export function parseDerivedFile(value: unknown, seeds: Seed[]): DerivedFile {
+  if (
+    !isRecord(value) ||
+    !isStringArray(value["fields"]) ||
+    !isRecord(value["derived"]) ||
+    !Object.values(value["derived"]).every(isDerivedEntry)
+  ) {
+    throw new Error("Invalid conformance derived fields");
+  }
+  const derivedFile = value as unknown as DerivedFile;
+
+  assertKeys("derived-fields.json fields", derivedFile.fields, DERIVED_KEYS);
+  const entries = Object.keys(derivedFile.derived);
+  if (entries.length !== seeds.length) {
+    throw new Error(
+      `derived-fields.json has ${entries.length} entries for ${seeds.length} seeds`,
+    );
+  }
+  for (const seed of seeds) {
+    const entry = derivedFile.derived[seed.id];
+    if (!entry) {
+      throw new Error(`derived-fields.json has no entry for seed ${seed.id}`);
+    }
+    assertKeys(
+      `derived-fields.json derived["${seed.id}"]`,
+      Object.keys(entry),
+      DERIVED_KEYS,
+    );
+  }
+
+  return derivedFile;
 }
 
 // -- the golden wire fixtures --------------------------------------------------------------------
